fixed a dup issue in the iconPostion enum
added in SF icon code, looks to all work. The animation settings are a switch based on an enum since animation support is currently "limited" to a smallish number of types. If that grows past a certain point, I'll remove the enum and make animation type just a normal string.